First working algorithm. Need improvements, and need to write a checker and the output to file part.

+++ src/AC2017Qualif/EndPoint.java
@@ -8,13 +8,13 @@ import java.util.TreeMap;
 public class EndPoint implements Comparable<EndPoint>{
 
 	public final int epID;
-	public final int LD;//the latency of serving a video request from the data center to this endpoint, in milliseconds
+	public final long LD;//the latency of serving a video request from the data center to this endpoint, in milliseconds
 	public final int K;//the number of cache servers that this endpoint is connected to
 	public ArrayList<Server> ServerList; // list of all servers
-	public final ArrayList<Integer> Latency4ServerList;// Latency to all servers. Put  Integer.MAX_VALUE for server not connected
+	public final ArrayList<Long> Latency4ServerList;// Latency to all servers. Put  Integer.MAX_VALUE for server not connected
 	public Map<Integer,Request> RequestList;// List of requests, mapped by index of video !!! Need concatenate if 2 requests are from same Endpoint with same video
 	
-	public EndPoint(int epID, int lD, int k, ArrayList<Integer>  latency4ServerList,ArrayList<Server> ServerList) {
+	public EndPoint(int epID, long lD, int k, ArrayList<Long>  latency4ServerList,ArrayList<Server> ServerList) {
 		super();
 		this.epID = epID;
